Enable API for search by keyword and moveid
