+++ api/app.js
@@ -4,7 +4,8 @@ const path = require('path');
 const db = require('./models');
 const app = express();
 const PORT = process.env.PORT;
-
+const API_KEY = process.env.API_KEY
+const https = require('https');
 
 // this lets us parse 'application/json' content in http requests
 app.use(express.json());
@@ -29,6 +30,43 @@ if(process.env.NODE_ENV==='production') {
 // update DB tables based on model updates. Does not handle renaming tables/columns
 // NOTE: toggling this to true drops all tables (including data)
 db.sequelize.sync({ force: false });
+
+app.get('/search/keyword/:keyword', (req, res) => {
+
+  https.get('https://api.themoviedb.org/3/search/movie?'+'api_key='+API_KEY+'&language=en-US&page=1&include_adult=false&query='+req.params.keyword, (response) => {
+    let data = '';
+
+    // A chunk of data has been received.
+    response.on('data', (chunk) => {
+      data += chunk;
+    });
+
+    // The whole response has been received. Print out the result.
+    response.on('end', () => {
+      //console.log(data)
+      res.send(JSON.parse(data).results);
+    });
+  }).on("error", (err) => { console.log("Error: " + err.message); });
+})
+
+
+app.get('/search/movie/:movieid', (req, res) => {
+
+  https.get('https://api.themoviedb.org/3/movie/'+req.params.movieid+'/watch/providers?api_key='+API_KEY, (response) => {
+    let data = '';
+
+    // A chunk of data has been received.
+    response.on('data', (chunk) => {
+      data += chunk;
+    });
+
+    // The whole response has been received. Print out the result.
+    response.on('end', () => {
+      //console.log(data)
+      res.send(JSON.parse(data).results.US.flatrate);
+    });
+  }).on("error", (err) => { console.log("Error: " + err.message); });
+})
 
 // start up the server
 if(PORT) {
